requirements.txt
discord.py[voice] >= 1.5.0
jishaku
aiohttp
koreanbots